DEV-2766 refactor client bootstrap
