+++ cmd/rport/cli/params_init.go
@@ -1,4 +1,4 @@
-package cliboilerplate
+package cli
 
 import (
 	"os"
