+++ cmd/rportd/main.go
@@ -11,7 +11,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/cloudradar-monitoring/rport/cmd/rportd/cliboilerplate"
+	"github.com/cloudradar-monitoring/rport/cmd/rportd/service_management"
 	"github.com/cloudradar-monitoring/rport/share/logger"
 
 	"github.com/kardianos/service"
@@ -446,7 +446,7 @@ func runMain(*cobra.Command, []string) {
 			}
 		}
 
-		err := cliboilerplate.HandleSvcCommand(*svcCommand, *cfgPath, svcUser)
+		err := service_management.HandleSvcCommand(*svcCommand, *cfgPath, svcUser)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -495,7 +495,7 @@ func runMain(*cobra.Command, []string) {
 	}
 
 	if !service.Interactive() {
-		err = cliboilerplate.RunAsService(s, *cfgPath)
+		err = service_management.RunAsService(s, *cfgPath)
 		if err != nil {
 			log.Fatal(err)
 		}
